+++ experimental/transformer/include/ops/W8A8B8O8Linear.h
@@ -1,0 +1,20 @@
+#include "common.h"
+
+struct W8A8B8O8Linear_params {
+    Matrix3D<int8_t> weight;
+    Matrix3D<int32_t> bias;
+    float alpha;
+    float beta;
+};
+
+class W8A8B8O8Linear{
+public:
+    W8A8B8O8Linear(W8A8B8O8Linear_params &params_);
+    W8A8B8O8Linear(){};
+    void forward(const Matrix3D<int8_t> &x, Matrix3D<int8_t> &output);
+    struct matmul_params params;
+    float alpha;
+    float beta;
+};
+
+void load_W8A8B8O8Linear_params(W8A8B8O8Linear &op, std::string prefix);
